♻️default-star系をlocal-starsから独立
